pyproject.toml
[tool.poetry]
name = "ocx-validator"
version = "1.3.0"
description = "Console application for the validation of 3Docx models against the OCX schema"
authors = ["Ole Christian Astrup <ole.christian.astrup@dnv.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://3docx.org"
repository = "https://github.com/OCXStandard/schema_validator.git"
keywords = ["click", "lxml", "lint"]
classifiers = [
    "Environment :: Console",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Quality Assurance",
    "Development Status :: 3 - Alpha",
]
include = [
    "LICENSE",
]


[tool.poetry.dependencies]
python = "^3.9"
click = "^8.0.3"
lxml = "^4.7.1"
colorlog = "^6.6.0"
PyYAML = "^6.0"
tabulate = "^0.8.9"
pylint = "^2.12.2"

[tool.poetry.dev-dependencies]
pytest = "^7.0.1"
pytest-cov = "^3.0.0"

# New scripts
[tool.poetry.scripts]
schema-validator = 'ocx-validator:main'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
